requirements.txt
fastapi~=0.115.0
uvicorn~=0.30.6
Scrapy~=2.11.2
itemadapter~=0.9.0
minio~=7.2.8
pydantic~=2.9.2
python-dotenv~=1.0.1